Clean-up helper.dart, extract to another files
